requirement.txt
amqp==2.2.2
billiard==3.5.0.3
cairocffi==0.8.0
CairoSVG==1.0.22
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cssselect2==0.2.1
Django==1.11.5
django-bootstrap3==7.1.0
django-celery-results==1.0.1
django-pg-utils==0.1.5
django-widget-tweaks==1.4.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
kombu==4.1.0
olefile==0.44
pdfrw==0.4
Pillow==4.2.1
psycopg2==2.7.3.1
pycparser==2.18
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.8.1
pytz==2017.2
reportlab==3.4.0
requests==2.18.4
six==1.11.0
sorl-thumbnail==12.4.1
tinycss2==0.6.1
Unipath==1.1
urllib3==1.22
vine==1.1.4
WeasyPrint==0.41
webencodings==0.5.1
xhtml2pdf==0.2b1
